+++ mobile/TSweetMediasCommunicator.m
@@ -22,13 +22,13 @@
 
 -(TSweetResponse *) like:(NSInteger)mediaId
 {
-    NSString * route = [NSString stringWithFormat:@"/medias/%d/like", mediaId];
+    NSString * route = [NSString stringWithFormat:@"/medias/%ld/like", (long)mediaId];
     return [[TSweetRest shared] get:route];
 }
 
 -(TSweetResponse *) comment:(NSInteger)mediaId comment:(NSString *)comment
 {
-    NSString * route = [NSString stringWithFormat:@"/medias/%d/comment", mediaId];
+    NSString * route = [NSString stringWithFormat:@"/medias/%ld/comment", (long)mediaId];
     NSDictionary * parameters = @{
                                   @"comment": comment
                                   };
@@ -38,7 +38,7 @@
 
 -(TSweetResponse *) likeComment:(NSInteger)mediaId commentId:(NSInteger)commentId
 {
-    NSString * route = [NSString stringWithFormat:@"/medias/%d/comments/%d", mediaId, commentId];
+    NSString * route = [NSString stringWithFormat:@"/medias/%ld/comments/%ld", (long)mediaId, (long)commentId];
     return [[TSweetRest shared] get:route];
 }
